pyproject.toml
[project]
name = "realtime-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.95",
	"uvicorn[standard]>=0.22",
	"motor>=3.1",
	"python-jose>=3.3",
	"passlib[bcrypt]>=1.7",
	"pydantic>=2.5",
	"pydantic-settings>=2.11.0",
	"python-dotenv>=1.2.1",
	"bcrypt==4.3.0",
	"python-socketio>=5.9",
]

